feat(sandbox): add simple retrieval chain
# Conflicts:
#	uv.lock
