+++ pyproject.toml
@@ -7,8 +7,10 @@ readme = "README.md"
 requires-python = "<3.13,>=3.12"
 
 dependencies = [
+    "datasets>=3.4.1",
     "esprima>=4.0.1",
     "google-genai>=1.4.0",
+    "grpcio==1.60.1",
     "langchain-postgres>=0.0.13",
     "langchain[community,huggingface,google-genai]>=0.3.20",
     "psycopg[binary]>=3.2.6",
